Add files via upload
Challenge - Air Pollution
Air Pollution is Rising!
Predicting Pollution Levels using Machine Learning
It is winter time in Delhi so Cody decided to go for a walk to the news stand and on reaching was surprised to see the air quality index of Delhi in the newspaper . So he decided to collect the air samples of different locations and then took these samples to his lab where he extracted five features of the air he collected which can be used to predict the air quality index and combined it with the air quality index given in the newspapers. You are provided with the data collected by Cody and your job is to design a machine learning model which given the features extracted by Cody can predict air quality. 

Data consist of two csv files :-
Train.csv
Testcsv
Train:- The train data consists of five feature columns and one target column.All the feature columns are numeric and target value is also numeric.

Test:- test data consist all the columns contained by train with the the exception of target column which u have to predict using the model trained on train data.

We will be predicting target by linear regression algorithm